// demo-ui/src/app/features/generated/reference/time/calendar/models/CalendarFullView.ts
import { CalendarForm } from './CalendarForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class CalendarFullView {

id: number;
form: CalendarForm;
canUpdate: boolean;

/* Specific Code Start */
/* Specific Code End */
}